moved minio stuff to minio service again
